feat(31) : 회원 승인 요청 조회, 승인, 거부(삭제) (#42)
* feat: 회원 승인 요청 조회, 승인, 삭제 구현

* refactor: dto에 id 추가, 승인/반려 검증 추가

* refactor: 승인 검증 수정

* refactor: 삭제 검증 예외 메시지 수정

* refactor: 의미없는 상수 제거

* feat : 휴가정보 초기화기능 구현

* refactor: 멤버 승인 http 메서드 post로 변경

* fix: mapper 수정

* refactor : 엔티티 타입변경에 따른 dto 타입변경

---------

+++ backend/team6/src/main/java/programmers/team6/domain/vacation/repository/JdbcVacationEligibilitiesRepository.java
@@ -47,7 +47,7 @@ public class JdbcVacationEligibilitiesRepository {
 			@Override
 			public void setValues(PreparedStatement ps, int i) throws SQLException {
 				VacationGrantInfo vacationGrantInfo = eligibilities.get(i);
-				ps.setInt(1, vacationGrantInfo.totalCount());
+				ps.setDouble(1, vacationGrantInfo.totalCount());
 				ps.setInt(2, vacationGrantInfo.vacationInfoId());
 				ps.setInt(3, vacationGrantInfo.version());
 			}
