+++ backend/team6/src/main/java/programmers/team6/domain/vacation/rule/vacationgranteligiblities/VacationGrantEligibility.java
@@ -3,7 +3,7 @@ package programmers.team6.domain.vacation.rule.vacationgranteligiblities;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
-public record VacationGrantEligibility(Integer id, LocalDateTime joinDate, Integer vacationCount, Integer version) {
+public record VacationGrantEligibility(Integer id, LocalDateTime joinDate, double vacationCount, Integer version) {
 	public LocalDate joinDateAsLocalDate() {
 		return joinDate.toLocalDate();
 	}
